// clients/typescript/src/proto/api/tsdb/tsdb.ts
export namespace tsdb {
// ...
    export namespace AggregationSpec {
        export enum AggregationFunc {
            NONE = 0,
            SUM = 1,
            COUNT = 2,
            AVERAGE = 3,
            MIN = 4,
            MAX = 5,
            FIRST = 6,
            LAST = 7
        }
    }
// ...
}